Fix tests

+++ apps/files_sharing/tests/sharedmount.php
@@ -49,8 +49,12 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 
 	protected function tearDown() {
 		if ($this->view) {
-			$this->view->unlink($this->folder);
-			$this->view->unlink($this->filename);
+			if ($this->view->file_exists($this->folder)) {
+				$this->view->unlink($this->folder);
+			}
+			if ($this->view->file_exists($this->filename)) {
+				$this->view->unlink($this->filename);
+			}
 		}
 
 		parent::tearDown();
@@ -59,59 +63,46 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 	/**
 	 * test if the mount point moves up if the parent folder no longer exists
 	 */
-	function testShareMountLoseParentFolder() {
+	public function testShareMountLoseParentFolder() {
 
 		// share to user
-		$fileinfo = $this->view->getFileInfo($this->folder);
-		$result = \OCP\Share::shareItem('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER,
-			self::TEST_FILES_SHARING_API_USER2, 31);
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_USER,
+			$this->folder,
+			self::TEST_FILES_SHARING_API_USER1,
+			self::TEST_FILES_SHARING_API_USER2,
+			\OCP\Constants::PERMISSION_ALL);
 
-		$statement = "UPDATE `*PREFIX*share` SET `file_target` = ? where `share_with` = ?";
-		$query = \OCP\DB::prepare($statement);
-		$arguments = array('/foo/bar' . $this->folder, self::TEST_FILES_SHARING_API_USER2);
-		$query->execute($arguments);
+		$share->setTarget('/foo/bar' . $this->folder);
+		$this->shareManager->moveShare($share, self::TEST_FILES_SHARING_API_USER2);
 
-		$query = \OCP\DB::prepare('SELECT * FROM `*PREFIX*share`');
-		$result = $query->execute();
-
-		$shares = $result->fetchAll();
-
-		$this->assertSame(1, count($shares));
-
-		$share = reset($shares);
-		$this->assertSame('/foo/bar' . $this->folder, $share['file_target']);
+		$share = $this->shareManager->getShareById($share->getFullId());
+		$this->assertSame('/foo/bar' . $this->folder, $share->getTarget());
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
-
 		// share should have moved up
 
-		$query = \OCP\DB::prepare('SELECT * FROM `*PREFIX*share`');
-		$result = $query->execute();
-
-		$shares = $result->fetchAll();
-
-		$this->assertSame(1, count($shares));
-
-		$share = reset($shares);
-		$this->assertSame($this->folder, $share['file_target']);
+		$share = $this->shareManager->getShareById($share->getFullId());
+		$this->assertSame($this->folder, $share->getTarget());
 
 		//cleanup
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
-		\OCP\Share::unshare('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER, self::TEST_FILES_SHARING_API_USER2);
+		$this->shareManager->deleteShare($share);
 		$this->view->unlink($this->folder);
 	}
 
 	/**
 	 * @medium
 	 */
-	function testDeleteParentOfMountPoint() {
-
+	public function testDeleteParentOfMountPoint() {
 		// share to user
-		$fileinfo = $this->view->getFileInfo($this->folder);
-		$result = \OCP\Share::shareItem('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER,
-			self::TEST_FILES_SHARING_API_USER2, 31);
-
-		$this->assertTrue($result);
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_USER,
+			$this->folder,
+			self::TEST_FILES_SHARING_API_USER1,
+			self::TEST_FILES_SHARING_API_USER2,
+			\OCP\Constants::PERMISSION_ALL
+		);
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 		$user2View = new \OC\Files\View('/' . self::TEST_FILES_SHARING_API_USER2 . '/files');
@@ -143,10 +134,14 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		$this->view->unlink($this->folder);
 	}
 
-	function testMoveSharedFile() {
-		$fileinfo = $this->view->getFileInfo($this->filename);
-		$result = \OCP\Share::shareItem('file', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER,
-			self::TEST_FILES_SHARING_API_USER2, 31);
+	public function testMoveSharedFile() {
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_USER,
+			$this->filename,
+			self::TEST_FILES_SHARING_API_USER1,
+			self::TEST_FILES_SHARING_API_USER2,
+			\OCP\Constants::PERMISSION_READ | \OCP\Constants::PERMISSION_UPDATE | \OCP\Constants::PERMISSION_SHARE
+		);
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 
@@ -166,22 +161,27 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		$this->assertTrue(\OC\Files\Filesystem::file_exists($this->filename));
 
 		//cleanup
-		\OCP\Share::unshare('file', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER, self::TEST_FILES_SHARING_API_USER2);
+		$this->shareManager->deleteShare($share);
 	}
 
 	/**
 	 * share file with a group if a user renames the file the filename should not change
 	 * for the other users
 	 */
-	function testMoveGroupShare () {
+	public function testMoveGroupShare () {
 		\OC_Group::createGroup('testGroup');
 		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
 		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
 		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
 
 		$fileinfo = $this->view->getFileInfo($this->filename);
-		$result = \OCP\Share::shareItem('file', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP,
-			"testGroup", 31);
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_GROUP,
+			$this->filename,
+			self::TEST_FILES_SHARING_API_USER1,
+			'testGroup',
+			\OCP\Constants::PERMISSION_READ | \OCP\Constants::PERMISSION_UPDATE | \OCP\Constants::PERMISSION_SHARE
+		);
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 
@@ -202,7 +202,7 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 
 		//cleanup
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
-		\OCP\Share::unshare('file', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP, 'testGroup');
+		$this->shareManager->deleteShare($share);
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
@@ -214,7 +214,7 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 	 * @param string $expectedResult
 	 * @param bool $exception if a exception is expected
 	 */
-	function testStripUserFilesPath($path, $expectedResult, $exception) {
+	public function testStripUserFilesPath($path, $expectedResult, $exception) {
 		$testClass = new DummyTestClassSharedMount(null, null);
 		try {
 			$result = $testClass->stripUserFilesPathDummy($path);
@@ -228,7 +228,7 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		}
 	}
 
-	function dataProviderTestStripUserFilesPath() {
+	public function dataProviderTestStripUserFilesPath() {
 		return array(
 			array('/user/files/foo.txt', '/foo.txt', false),
 			array('/user/files/folder/foo.txt', '/folder/foo.txt', false),
@@ -239,7 +239,7 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		);
 	}
 
-	function dataPermissionMovedGroupShare() {
+	public function dataPermissionMovedGroupShare() {
 		$data = [];
 
 		$powerset = function($permissions) {
@@ -256,7 +256,6 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		//Generate file permissions
 		$permissions = [
 			\OCP\Constants::PERMISSION_UPDATE,
-			\OCP\Constants::PERMISSION_CREATE,
 			\OCP\Constants::PERMISSION_SHARE,
 		];
 
@@ -321,54 +320,45 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
 
 		// Share item with group
-		$fileinfo = $this->view->getFileInfo($path);
-		$this->assertTrue(
-			\OCP\Share::shareItem($type, $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP,	"testGroup", $beforePerm)
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_GROUP,
+			$path,
+			self::TEST_FILES_SHARING_API_USER1,
+			'testGroup',
+			$beforePerm
 		);
 
 		// Login as user 2 and verify the item exists
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 		$this->assertTrue(\OC\Files\Filesystem::file_exists($path));
-		$result = \OCP\Share::getItemSharedWithBySource($type, $fileinfo['fileid']);
-		$this->assertNotEmpty($result);
-		$this->assertEquals($beforePerm, $result['permissions']);
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
+		$this->assertEquals($beforePerm, $result->getPermissions());
 
 		// Now move the item forcing a new entry in the share table
 		\OC\Files\Filesystem::rename($path, "newPath");
 		$this->assertTrue(\OC\Files\Filesystem::file_exists('newPath'));
 		$this->assertFalse(\OC\Files\Filesystem::file_exists($path));
 
-		// Login as user 1 again and change permissions
-		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
-		$this->assertTrue(
-			\OCP\Share::setPermissions($type, $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP, "testGroup", $afterPerm)
-		);
+		// change permissions
+		$share->setPermissions($afterPerm);
+		$this->shareManager->updateShare($share);
 
 		// Login as user 3 and verify that the permissions are changed
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER3);
-		$result = \OCP\Share::getItemSharedWithBySource($type, $fileinfo['fileid']);
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER3);
 		$this->assertNotEmpty($result);
-		$this->assertEquals($afterPerm, $result['permissions']);
-		$groupShareId = $result['id'];
+		$this->assertEquals($afterPerm, $result->getPermissions());
 
 		// Login as user 2 and verify that the permissions are changed
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
-		$result = \OCP\Share::getItemSharedWithBySource($type, $fileinfo['fileid']);
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
 		$this->assertNotEmpty($result);
-		$this->assertEquals($afterPerm, $result['permissions']);
-		$this->assertNotEquals($groupShareId, $result['id']);
-
-		// Also verify in the DB
-		$statement = "SELECT `permissions` FROM `*PREFIX*share` WHERE `id`=?";
-		$query = \OCP\DB::prepare($statement);
-		$result = $query->execute([$result['id']]);
-		$shares = $result->fetchAll();
-		$this->assertCount(1, $shares);
-		$this->assertEquals($afterPerm, $shares[0]['permissions']);
+		$this->assertEquals($afterPerm, $result->getPermissions());
+		$this->assertEquals('/newPath', $result->getTarget());
 
 		//cleanup
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
-		\OCP\Share::unshare($type, $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP, 'testGroup');
+		$this->shareManager->deleteShare($share);
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
@@ -388,66 +378,44 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 
 		// Share item with group
 		$fileinfo = $this->view->getFileInfo($this->folder);
-		$this->assertTrue(
-			\OCP\Share::shareItem('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP,	"testGroup", \OCP\Constants::PERMISSION_READ)
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_GROUP,
+			$this->folder,
+			self::TEST_FILES_SHARING_API_USER1,
+			'testGroup',
+			\OCP\Constants::PERMISSION_READ
 		);
 
 		// Login as user 2 and verify the item exists
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 		$this->assertTrue(\OC\Files\Filesystem::file_exists($this->folder));
-		$result = \OCP\Share::getItemSharedWithBySource('folder', $fileinfo['fileid']);
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
 		$this->assertNotEmpty($result);
-		$this->assertEquals(\OCP\Constants::PERMISSION_READ, $result['permissions']);
+		$this->assertEquals(\OCP\Constants::PERMISSION_READ, $result->getPermissions());
 
 		// Delete the share
 		$this->assertTrue(\OC\Files\Filesystem::rmdir($this->folder));
 		$this->assertFalse(\OC\Files\Filesystem::file_exists($this->folder));
 
 		// Verify we do not get a share
-		$result = \OCP\Share::getItemSharedWithBySource('folder', $fileinfo['fileid']);
-		$this->assertEmpty($result);
-
-		// Verify that the permission is correct in the DB
-		$qb = $connection->getQueryBuilder();
-		$qb->select('*')
-			->from('share')
-			->where($qb->expr()->eq('file_source', $qb->createParameter('fileSource')))
-			->andWhere($qb->expr()->eq('share_type', $qb->createParameter('shareType')))
-			->setParameter(':fileSource', $fileinfo['fileid'])
-			->setParameter(':shareType', 2);
-		$res = $qb->execute()->fetchAll();
-
-		$this->assertCount(1, $res);
-		$this->assertEquals(0, $res[0]['permissions']);
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
+		$this->assertEquals(0, $result->getPermissions());
 
 		// Login as user 1 again and change permissions
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
-		$this->assertTrue(
-			\OCP\Share::setPermissions('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP, "testGroup", \OCP\Constants::PERMISSION_ALL)
-		);
+		$share->setPermissions(\OCP\Constants::PERMISSION_ALL);
+		$share = $this->shareManager->updateShare($share);
 
 		// Login as user 2 and verify 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 		$this->assertFalse(\OC\Files\Filesystem::file_exists($this->folder));
-		$result = \OCP\Share::getItemSharedWithBySource('folder', $fileinfo['fileid']);
-		$this->assertEmpty($result);
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
+		$this->assertEquals(0, $result->getPermissions());
 
-		$connection = \OC::$server->getDatabaseConnection();
-		$qb = $connection->getQueryBuilder();
-		$qb->select('*')
-			->from('share')
-			->where($qb->expr()->eq('file_source', $qb->createParameter('fileSource')))
-			->andWhere($qb->expr()->eq('share_type', $qb->createParameter('shareType')))
-			->setParameter(':fileSource', $fileinfo['fileid'])
-			->setParameter(':shareType', 2);
-		$res = $qb->execute()->fetchAll();
-
-		$this->assertCount(1, $res);
-		$this->assertEquals(0, $res[0]['permissions']);
+		$this->shareManager->deleteShare($share);
 
 		//cleanup
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
-		\OCP\Share::unshare('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
